+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 All notable changes to the extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## [Unreleased]
+
+- Add settings to configure DataArray HTML representation verbosity:
+  `erlab.dataArray.displayExpandAttrs`, `erlab.dataArray.displayExpandCoords`, and
+  `erlab.dataArray.displayExpandData`.
+
+## [v0.2.0] - 2026-01-12
 
 - Add DataArrays and DataArray Detail panels for browsing and inspecting DataArrays.
 - Enrich hover actions to open details, pin, watch, and open ImageTool.